requirements.txt
discord.py
spotipy
python-dotenv
discord.py[voice]
PyNaCl
